qemu: Add luks support for domain disk
Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1301021

Generate the luks command line using the AES secret key to encrypt the
luks secret. A luks secret object will be in addition to a an AES secret.

For hotplug, check if the encinfo exists and if so, add the AES secret
for the passphrase for the secret object used to decrypt the device.

Modify/augment the fakeSecret* in qemuxml2argvtest in order to handle
find a uuid or a volume usage with a specific path prefix in the XML
(corresponds to the already generated XML tests). Add error message
when the 'usageID' is not 'mycluster_myname'. Commit id '1d632c39'
altered the error message generation to rely on the errors from the
secret_driver (or it's faked replacement).

Add the .args output for adding the LUKS disk to the domain

Signed-off-by: John Ferlan <jferlan@redhat.com>
(cherry picked from commit da86c6c22674ccc147224afa2740e33d8cbdbf22)

NB: The .args output was modified from upstream since the downstream
    does not contain commit id 'e114b09157b7fcca12b218b531debfbc0c3a09d7'
    which adds ",sockets=1,cores=1,threads=1" to the "-smp 1" on the
    command line

Signed-off-by: John Ferlan <jferlan@redhat.com>