Introduce SMM feature
https://bugzilla.redhat.com/show_bug.cgi?id=1304483

Since its release of 2.4.0 qemu is able to enable System
Management Module in the firmware, or disable it. We should
expose this capability in the XML. Unfortunately, there's no good
way to determine whether the binary we are talking to supports
it. I mean, if qemu's run with real machine type, the smm
attribute can be seen in 'qom-list /machine' output. But it's not
there when qemu's run with -M none. Therefore we're stuck with
version based check.

Also, for the RHEL-7.3 I had to adjust the machine-smm-opt test.
In the upstream we already retired QEMU_CAPS_SMP_TOPOLOGY
capability, but not in RHEL yet.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
(cherry picked from commit d0e4be9d0297494ba0e23843593f734234ddb08f)
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>

 Conflicts:
   tests/qemucapabilitiesdata/caps_2.6.0-gicv2.aarch64.xml
   tests/qemucapabilitiesdata/caps_2.6.0-gicv3.aarch64.xml
   tests/qemucapabilitiesdata/caps_2.6.0.ppc64le.xml -
   QEMU_CAPS_DISPLAY is not backported yet

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>