conf: report an error message for non-existing USB hubs
If any of the devices referenced a USB hub that does not exist,
defining the domain would either fail with:
error: An error occurred, but the cause is unknown
(if only the last hub in the path is missing)
or crash.

Return a proper error instead of crashing.

https://bugzilla.redhat.com/show_bug.cgi?id=1367130
(cherry picked from commit ef66bd5df8af93db5a77f0c1f33182139c9b63b3)
Signed-off-by: Ján Tomko <jtomko@redhat.com>