Revert "admin: Fix the default uri for session daemon to libvirtd:///session"
The problem is, that the patch fixed the issue in its own way because the
original upstream commit 30ce2f0e (which was an incorrect upstream fix anyway)
was dependent on commit f5da0d18 which wasn't backported, thus causing
conflicts. However, it turned out that f5da0d18 had to be backported
eventually, therefore the issue with session daemon uri for virt-admin can be
fixed by a proper backport of upstream commit 9e5e7f3a (that fixes 30ce2f0e in
upstream).

https://bugzilla.redhat.com/show_bug.cgi?id=1367269
Signed-off-by: Erik Skultety <eskultet@redhat.com>