conf: Provide error on undefined iothreadsched entry
https://bugzilla.redhat.com/show_bug.cgi?id=1366484

When commit id '6dfb4507' refactored where the iothreadsched data was
stored, the error message for when the virDomainIOThreadIDFind failed
to find an iothreadid ("iothreadsched attribute 'iothreads' uses
undefined iothread ids") was lost. This led to the possibility that
someone would try to use it, but receive the generic message "An error
occurred, but the cause is unknown".

This patch adds the error message back so that someone will know that
they have an invalid configuration.

Signed-off-by: John Ferlan <jferlan@redhat.com>
(cherry picked from commit 8c6b29e7b41d678ab9c5542efd607b28776e4f24)