cpu_x86: Fix minimum match custom CPUs on hosts with CMT
Since the introduction of CMT features (commit v1.3.5-461-gf294b83)
starting a domain with custom CPU and match='minimum' on a host which
supports CMT fails because QEMU complains about unknown 'cmt' feature.

"cpu_x86: Fix host-model CPUs on hosts with CMT" commit fixed similar
issue for host-model CPUs.

This patch is a RHEL-only hack because upstream fixes this by unifying
the code for these two types of CPUs, but the upstream solution is
invasive and cannot be easily backported.

https://bugzilla.redhat.com/show_bug.cgi?id=1365500

Signed-off-by: Jiri Denemark <jdenemar@redhat.com>