qemu: Add support for using AES secret for SCSI hotplug
https://bugzilla.redhat.com/show_bug.cgi?id=1406442

Support for virtio disks was added in commit id 'fceeeda', but not for
SCSI drives. Add the secret for the server when hotplugging a SCSI drive.
No need to make any adjustments for unplug since that's handled during
the qemuDomainDetachDiskDevice call to qemuDomainRemoveDiskDevice in
the qemuDomainDetachDeviceDiskLive switch.

Added a test to/for the command line processing to show the command line
options when adding a SCSI drive for the guest.

(cherry picked from commit 0701abcb3ba78ba27cf1f47e01b3d9607ad37b72)

Resolved conflict since upstream commit id '97ca6eed9a' is not present
(just remove the drivealias changes)

https://bugzilla.redhat.com/show_bug.cgi?id=1411398

Signed-off-by: John Ferlan <jferlan@redhat.com>