Increase default file handle limits for daemons
Linux still defaults to a 1024 open file handle limit. This causes
scalability problems for libvirtd / virtlockd / virtlogd on large
hosts which might want > 1024 guest to be running. In fact if each
guest needs > 1 FD, we can't even get to 500 guests. This is not
good enough when we see machines with 100's of physical cores and
TBs of RAM.

In comparison to other memory requirements of libvirtd & related
daemons, the resource usage associated with open file handles
is essentially line noise. It is thus reasonable to increase the
limits unconditionally for all installs.

Resolves: https://bugzilla.redhat.com/1442043 (RHEL7.3.z)
Resolves: https://bugzilla.redhat.com/1429551 (RHEL7.4)

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>
(cherry picked from commit 27cd76350021d36b9bd8b187ce5c8919659e3806)