Ignore Frodo messages when parsing log context
Libvirt traditionally spawns a Qemu process and monitors its output (to
stdout/stderr), redirecting such messages to the VM's logfile. On
various circumstances, libvirt reads this logfile to pickup unexpected
error message and/or output from Qemu.

On AHV, libvirt instead spawns a wrapper which, in turn, spawns Qemu and
Frodo. All three components (Frodo, Qemu and the Wrapper) now log to
stdout/stderr. This breaks libvirt's capabilities of understanding what
has happened to a Qemu that failed (eg. ENG-104920).

Furtunately, libvirt already has a facility to filter out its own
messages written to this file. Let's piggy-back on it and tell it to
also ignore everything that contains "frodo".

Change-Id: Ifcc4aa1899b4694cd2500827defb867932865b61