ENG-106003: Grab unbindLock before unbinding nvidia GPU.
Add a special case to grab an nvidia unbindLock before unbinding
an nvidia PCI device.

This solution is not ideal. Preferably we should have a hook
mechanism to do whatever is necessary to prepare a device in general
for unbinding, but this is significantly more work to add this
mechanism that would be carried forward as a patch we'd need to
maintain against upstream.

Change-Id: I88d19649cadb4b41f287e67f6cc1e10ecee03dd1